Added default player idle state and calls it on Game managers load state. Capture Bullet no longer collide with cleared bubbles.

+++ Assets/Scripts/Player/CaptureBulletBehavior.cs
@@ -19,10 +19,13 @@ public class CaptureBulletBehavior : MonoBehaviour
         if (disabled) return;
         switch (collision.gameObject.tag)
         {
-            case "BubbleSpirit":               
-                BubbleSpirit capturedBubble = collision.GetComponent<BubbleSpirit>();
-                ParentPlayer.SetCapture(capturedBubble);               
-                destroySelf();
+            case "BubbleSpirit":
+                if (collision.GetComponent<BubbleSpirit>().state == BubbleSpirit.State.NORMAL)
+                {
+                    BubbleSpirit capturedBubble = collision.GetComponent<BubbleSpirit>();
+                    ParentPlayer.SetCapture(capturedBubble);
+                    destroySelf();
+                }
                 break;
             case "Wall Top":
                 disabled = true;
